# Add and default to `DebugOpt` build type for faster PRs