This PR adds the new value `DebugOpt` for the CMAKE_BUILD_TYPE variable. It contains the flags "-O2 -g". The difference to the existing `RelWithDebInfo` is that `-DNDEBUG` is *not* part of `DebugOpt`. It thus keeps all internal assertions, but compiles with -O2. Also switch the default from `Debug` to `DebugOpt` for the CI.

Locally, I go from Debug:
```
time ninja -j16
real	4m22,021s
user	51m27,560s
sys	2m55,290s

time ctest -j15
real	19m36,899s
user	257m17,031s
sys	5m50,710s
```
to DebugOpt:
```
time ninja -j16
real	5m48,021s
user	74m6,785s
sys	3m25,686s

time ctest -j15
real	9m49,003s
user	122m58,941s
sys	5m4,590s
```
My hope is that the CI will be faster. It might also result in some -Werrors since optimization enables more of the compiler's checks and/or in some failed test cases. I would like to ask for help in fixing those in code I'm not too familiar with.