@fbrausse 2023-10-18
@rafaelsamenezes Could you take a look?
```
In file included from /home/runner/work/esbmc/esbmc/unit/goto-programs/interval_template.test.cpp:1:
/home/runner/work/esbmc/esbmc/src/goto-programs/abstract-interpretation/interval_template.h: In function ‘interval_templatet<int> operator*(const interval_templatet<int>&, const interval_templatet<int>&)’:
/home/runner/work/esbmc/esbmc/src/goto-programs/abstract-interpretation/interval_template.h:18:7: error: ‘result.interval_templatet<int>::lower’ may be used uninitialized [-Werror=maybe-uninitialized]
   18 | class interval_templatet
      |       ^~~~~~~~~~~~~~~~~~
/home/runner/work/esbmc/esbmc/src/goto-programs/abstract-interpretation/interval_template.h:343:27: note: ‘result.interval_templatet<int>::lower’ was declared here
  343 |     interval_templatet<T> result;
      |                           ^~~~~~
```
https://github.com/esbmc/esbmc/actions/runs/6564879380/job/17832190303?pr=1403