@fbrausse 2023-10-19
Thanks, Rafael.

> Once this is merged, it would be good to extend the windows script to use this.

I agree. If it's easier, I can push this branch to this repository so we can all push commits there. Would probably need to open a new PR, though.