@fbrausse 2023-10-19
@rafaelsamenezes I think the maybe_uninitialized warning is not a false positive but actually happens when the synthesized copy-ctor copies `lower` even though `lower_set` is false. It should be possible to avoid that one by combining both into a member of type `std::optional<T>` (which internally is supposed to properly handle initialization of the storage associated with T). Though that may be a too intrusive change of the API.