@rafaelsamenezes 2023-10-19
> t should be possible to avoid that one by combining both into a member of type std::optional<T> (which internally is supposed to properly handle initialization of the storage associated with T). Though that may be a too intrusive change of the API.

I like this change actually.