@rafaelsamenezes 2023-10-20
@fbrausse could you rebase and check whether the Intervals are still giving an error?