@fbrausse 2023-10-20
For the Ubuntu CI we now have ~29m build and ~41m running tests. That's much better than before (#1407): ~21m build and 1h 24m running tests.